+++ apps/base/models/__init__.py
@@ -1,0 +1,6 @@
+from apps.base.models.base import BaseModel
+
+__all__ = (
+    'BaseModel',
+)
+
