+++ apps/base/models/base.py
@@ -1,0 +1,12 @@
+from django.db import models
+import uuid
+
+
+class BaseModel(models.Model):
+
+    id = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)
+    date_created = models.DateTimeField(auto_now_add=True, editable=False)
+    date_modified = models.DateTimeField(auto_now=True, editable=False)
+
+    class Meta: 
+        abstract = True
