+++ apps/notes/models/note.py
@@ -1,0 +1,13 @@
+from django.db import models
+from apps.base.models import BaseModel
+from apps.user.models import User
+
+
+class Note(BaseModel):
+    
+    body = models.TextField()
+    author = models.ForeignKey(User, on_delete=models.CASCADE, related_name='blog_posts')
+
+    class Meta:
+        db_table = 'note'
+        verbose_name = 'note'
