+++ apps/team/models/__init__.py
@@ -1,0 +1,7 @@
+from .team import Team
+from .user_team import UserTeam
+
+__all__ = (
+    'Team',
+    'UserTeam'
+)
