+++ apps/user/models/__init__.py
@@ -1,0 +1,7 @@
+from apps.user.models.user import User
+
+__all__ = (
+    'User',
+)
+
+
