--- team/models/__init__.py
@@ -1,4 +1,0 @@
-from .team import Team as _Team
-from .userteam import UserTeamRelation as _UserTeamRelation
-
-__all__ = ['Team', 'UserTeamRelation']
