--- user/models.py
@@ -1,20 +1,0 @@
-from django.db import models
-from django.contrib.auth.models import BaseUserManager, AbstractBaseUser, AbstractUser, PermissionsMixin
-import uuid
-
-# Create your models here.
-
-class User (AbstractUser):
-
-    id = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)
-    date_created = models.DateTimeField(auto_now_add=True, editable=False)
-    date_modified = models.DateTimeField(auto_now=True, editable=False)
-    username = models.CharField(unique=True, max_length=255)
-    first_name = models.CharField(max_length=255)
-    last_name = models.CharField(max_length=255)
-
-    USERNAME_FIELD = 'username'
-
-    class Meta:
-        db_table = 'user_table'
-        verbose_name = "user"
